requirements.txt
requests>=2.0.0
PyYAML>=5.0.0
setuptools>=68.1.2